added fb profile image
