+++ src/AuthContext.js
@@ -3,13 +3,23 @@ import {
   createUserWithEmailAndPassword,
   signInWithEmailAndPassword,
   onAuthStateChanged,
+  GoogleAuthProvider,
+  FacebookAuthProvider,
+  signInWithPopup,
   signOut,
 } from 'firebase/auth';
+import { useNavigate } from 'react-router-dom';
+import { updateProfile } from '@firebase/auth';
 import { auth } from './firebase';
 
 const userAuthContext = createContext();
+const googleProvider = new GoogleAuthProvider();
+const fbProvider = new FacebookAuthProvider();
+
 
 export function UserAuthContextProvider({ children }) {
+  const navigate = useNavigate();
+
   const [user, setUser] = useState({});
 
   async function logIn(email, password) {
@@ -20,6 +30,33 @@ export function UserAuthContextProvider({ children }) {
     return createUserWithEmailAndPassword(auth, email, password);
   }
 
+  const googleLogin = async () => {
+    try {
+      const result = await signInWithPopup(auth, googleProvider);
+      navigate('/');
+      return result;
+    } catch (error) {
+      return { error: error.message };
+    }
+  };
+
+  const fbLogin = async () => {
+    try {
+      const result = await signInWithPopup(auth, fbProvider);
+      const credentials = await FacebookAuthProvider.credentialFromResult(
+        result
+      );
+      const token = credentials.accessToken;
+      const photoUrl = `${result.user.photoURL}?access_token=${token}`;
+      await updateProfile(auth.currentUser, { photoURL: photoUrl });
+      // eslint-disable-next-line
+      console.log(token);
+      navigate('/');
+      return result;
+    } catch (error) {
+      return { error: error.message };
+    }
+  };
   async function logOut() {
     return signOut(auth);
   }
@@ -37,7 +74,7 @@ export function UserAuthContextProvider({ children }) {
   }, []);
 
   const methods = useMemo(
-    () => ({ loggedInUser: user, logIn, signUp, logOut }),
+    () => ({ loggedInUser: user, logIn, signUp, logOut, googleLogin, fbLogin }),
     []
   );
   return (
